Introduce golos.worker.hpp file
